+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Constraints` type now conforms to `Collection` protocol (backed by `Array`)
 - Add `clamp(to limit: ClosedRange<CGFloat>)` API for dimension anchors
+- Add `Constraints` `activate()` and `deactivate()` methods
 
 ## Align 2.2.1